moving things around to integrate in bmcmt
