+++ tla-bmcmt/pom.xml
@@ -40,6 +40,11 @@
             <artifactId>tla-assignments</artifactId>
             <version>${project.version}</version>
         </dependency>
+        <dependency>
+            <groupId>at.forsyte.apalache</groupId>
+            <artifactId>tla-types</artifactId>
+            <version>${project.version}</version>
+        </dependency>
 
         <!-- parent's dependencies -->
         <dependency>
